// src/app/shared/models/bank.model.ts
export class Bank {
  name: string = '';
  account_number: string = '';
  bank_name: string = '';
  bank_code: string = '';
  routing_code: string = '';
  branch: string = '';
  routing_method: string = '';
  swift_code: string = '';
  phone_number: string = '';
  country_code: string = '';
  currency_id: string = '';
  pay_type: string = '';
  ifsc_code: string = '';
}